32 bit repr

+++ src/lib.rs
@@ -116,7 +116,7 @@ impl BitmapHints {
     ///
     /// If there are no offset present at that height, aka an overflow, or the entry has already
     /// been fetched.
-    pub fn get_indexes(&mut self, height: BlockHeight) -> Vec<u16> {
+    pub fn get_indexes(&mut self, height: BlockHeight) -> Vec<u32> {
         let file_pos = self
             .map
             .get(&height)
@@ -138,7 +138,7 @@ impl BitmapHints {
                 curr_byte = u8::from_le_bytes(single_byte_arr);
             }
             if ((curr_byte >> leftovers) & 0x01) == 0x01 {
-                unspents.push(bit_pos as u16);
+                unspents.push(bit_pos as u32);
             }
         }
         unspents
